tests/submission/PhilipTaylor/canvas/2d.type.exists.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.type.exists</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.type.html">type</a>.exists</h1>
<p class="desc">The 2D context interface is a property of 'window'</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.2d.type">context.2d.type</a>
</li></ul>
</div>
<p class="notes">Defined in "Web IDL" (draft)
</p><p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("The 2D context interface is a property of 'window'");
_addTest(function(canvas, ctx) {

_assert(window.CanvasRenderingContext2D, "window.CanvasRenderingContext2D");


});
</script>

<div id="log"></div>
</body></html>
